Portability fix for OpenBSD in time quota external acl helper.

+++ helpers/external_acl/time_quota/ext_time_quota_acl.cc
@@ -96,7 +96,8 @@ static void vlog(const char *level, const char *format, va_list args)
 {
     time_t now = time(NULL);
 
-    fprintf(logfile, "%ld %s| %s: ", now, program_name, level);
+    fprintf(logfile, "%ld %s| %s: ", static_cast<long int>(now), 
+              program_name, level);
     vfprintf (logfile, format, args);
     fflush(logfile);
 }
